Release 1.0.1: Added interface to the controller and few code cleanup.
