+++ changelog.md
@@ -2,6 +2,10 @@ Changelog
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 2024-03-05
+### Added
+- Added interface for the controller.
+
 ## [1.0.0] - 2024-03-05
 ### Added
 - Initial release of the module.
