+++ pom.xml
@@ -10,7 +10,7 @@
 	</parent>
 	<groupId>com.modakdev</groupId>
 	<artifactId>cameroscopy</artifactId>
-	<version>1.0.0-SNAPSHOT</version>
+	<version>1.0.1-SNAPSHOT</version>
 	<name>cameroscopy</name>
 	<description>Spring boot logon module</description>
 	<properties>
