+++ src/main/java/com/modakdev/cameroscopy/repo/CameroscopyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CameroscopyRepository extends JpaRepository<CameroscopyUser, Long> {
 
-    public CameroscopyUser deleteUserByID(Long id);
+    public CameroscopyUser deleteUserById(Long id);
 }
