+++ src/main/resources/application.properties
@@ -14,7 +14,7 @@ logging.level.org.springframework.security=TRACE
 
 
 
-spring.main.allow-circular-references=true
+#spring.main.allow-circular-references=true
 
 
 server.servlet.context-path=/cameroscopy
